refactor: extract bootstrap seeder and seed data
- core/bootstrap.py logic → bootstrap/seeder.py
- SEED_SUPPLIERS + SEED_PRODUCTS dicts → bootstrap/seed_data.py
- Delete core/bootstrap.py
- Update main.py lifespan import
- Update test patches
